# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "GooseTrack",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://your-tasks-hv5t.onrender.com/",
      "description": "Personal to-do list website"
    }
  ],
  "paths": {
...
    "/tasks": {
      "get": {
        "summary": "Get tasks collection for a month",
        "tags": ["Task"],
        "operationId": "getAllTasks",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "default": "Successfully"
                    },
                    "tasks": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Task"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["message"],
                  "properties": {
                    "message": {
                      "type": "string",
                      "default": "Choose another type of language. Available options: ua, en."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["message"],
                  "properties": {
                    "message": {
                      "type": "string",
                      "default": "Not authorized"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ServerError" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create new task",
        "tags": ["Task"],
        "operationId": "addTask",
        "requestBody": {
          "description": "The new task object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "en": {
                      "$ref": "#/components/schemas/Task/properties/title/properties/en"
                    }
                  },
                  "start": {
                    "$ref": "#/components/schemas/Task/properties/start"
                  },
                  "end": { "$ref": "#/components/schemas/Task/properties/end" },
                  "date": {
                    "$ref": "#/components/schemas/Task/properties/date"
                  },
                  "status": {
                    "$ref": "#/components/schemas/Task/properties/status"
                  },
                  "priority": {
                    "$ref": "#/components/schemas/Task/properties/priority"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "The token issued to the current user",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "default": "Successfully"
                    },
                    "taskData": {
                      "type": "object",
                      "properties": {
                        "title": {
                          "en": {
                            "$ref": "#/components/schemas/Task/properties/title/properties/en"
                          }
                        },
                        "start": {
                          "$ref": "#/components/schemas/Task/properties/start"
                        },
                        "end": {
                          "$ref": "#/components/schemas/Task/properties/end"
                        },
                        "date": {
                          "$ref": "#/components/schemas/Task/properties/date"
                        },
                        "status": {
                          "$ref": "#/components/schemas/Task/properties/status"
                        },
                        "priority": {
                          "$ref": "#/components/schemas/Task/properties/priority"
                        },
                        "owner": {
                          "$ref": "#/components/schemas/Task/properties/owner"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["message"],
                  "properties": {
                    "message": {
                      "type": "string",
                      "default": "Not authorized"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["message"],
                  "properties": {
                    "message": {
                      "type": "string",
                      "default": "Server Error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Task": {
        "type": "object",
        "required": [
          "title",
          "start",
          "end",
          "date",
          "status",
          "priority",
          "owner"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "example": "642fdf8745bc73b2aecbae7b",
            "description": "Backend-generated unique identifier."
          },
          "title": {
            "type": "object",
            "properties": {
              "en": {
                "type": "string",
                "description": "Task title in English",
                "example": "Drink coffee"
              },
              "ua": {
                "type": "string",
                "description": "Task title in Ukrainian",
                "example": "Випити кави"
              }
            }
          },
          "start": {
            "type": "string",
            "format": "time",
            "example": "07:28",
            "description": "Tasks start"
          },
          "end": {
            "type": "string",
            "format": "time",
            "example": "07:28",
            "description": "Tasks end"
          },
          "date": {
            "type": "string",
            "format": "date",
            "example": "2023-04-12",
            "description": "Time of task start"
          },
          "status": {
            "type": "string",
            "enum": ["todo", "inprogress", "done"],
            "default": "todo",
            "description": "Task status"
          },
          "priority": {
            "type": "string",
            "enum": ["low", "medium", "high"],
            "default": "low",
            "description": "Task priority"
          },
          "owner": {
            "type": "string",
            "example": "642fdf8745bc73b2aecbae7b",
            "description": "Owner's unique identifier."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-04-12T07:28:49.493Z",
            "description": "Backend-generated date and time of task creating"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-04-12T07:28:49.493Z",
            "description": "Backend-generated date and time of task updating"
          }
        }
      },
      "ServerError": {
        "description": "Internal Server Error",
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": {
            "type": "string",
            "default": "Server Error"
          }
        }
      },
...
    }
  }
}